// src/models/search-setting.ts
export class SearchSettingBy{
    id: number;
    appCode: string;
    displayName: string;
    name: string;
    data: string;
    status:number;
    startDate: string;
    endDate: string;
    take: number;
    skip: number;
    names:string[];
    createdBy: number;
    createdDate: string;
    updatedBy: number;
    updatedDate: string;
    deletedBy: number;
    deletedDate: string;
    nameLike: string;
    appCodesIn:string[];
}